Refactored stumped and mstumped tests
