+++ .flake8
@@ -12,9 +12,9 @@ ignore =
 max-line-length = 88
 docstring-convention=numpy
 exclude = 
-    tests/*,
     docs,
     setup.py
 per-file-ignores =
     stumpy/__init__.py:D104
     stumpy/stumpi.py:E743
+    tests/*:D101,D102,D103
