+++ tests/test_floss.py
@@ -2,7 +2,6 @@ import numpy as np
 import numpy.testing as npt
 from stumpy import fluss, stump, aamp, core, floss
 from stumpy.floss import _nnmark, _iac, _cac, _rea
-import copy
 import pytest
 import naive
