+++ tests/test_scrump.py
@@ -1,6 +1,5 @@
 import numpy as np
 import numpy.testing as npt
-import pandas as pd
 from stumpy import scrump, core, stump, config
 from stumpy.scrump import prescrump
 import pytest
@@ -28,7 +27,6 @@ def naive_prescrump(T_A, m, T_B, s, exclusion_zone=None):
     distance_matrix = naive.distance_matrix(T_A, T_B, m)
 
     n_A = T_A.shape[0]
-    n_B = T_B.shape[0]
     l = n_A - m + 1
 
     P = np.empty(l)
@@ -281,7 +279,7 @@ def test_scrump_A_B_join_swap(T_A, T_B, percentages):
         np.random.seed(seed)
         ref_mp = naive_scrump(T_B, m, T_A, percentage, None, False, None)
         ref_P = ref_mp[:, 0]
-        ref_I = ref_mp[:, 1]
+        # ref_I = ref_mp[:, 1]
         ref_left_I = ref_mp[:, 2]
         ref_right_I = ref_mp[:, 3]
 
@@ -291,7 +289,7 @@ def test_scrump_A_B_join_swap(T_A, T_B, percentages):
         )
         approx.update()
         comp_P = approx.P_
-        comp_I = approx.I_
+        # comp_I = approx.I_
         comp_left_I = approx.left_I_
         comp_right_I = approx.right_I_
 
@@ -382,7 +380,6 @@ def test_scrump_self_join_full(T_A, T_B):
 def test_scrump_A_B_join_full(T_A, T_B):
 
     m = 3
-    zone = int(np.ceil(m / 4))
 
     ref_mp = naive.stamp(T_A, m, T_B=T_B)
     ref_P = ref_mp[:, 0]
@@ -421,7 +418,6 @@ def test_scrump_A_B_join_full(T_A, T_B):
 def test_scrump_A_B_join_full_swap(T_A, T_B):
 
     m = 3
-    zone = int(np.ceil(m / 4))
 
     ref_mp = naive.stamp(T_B, m, T_B=T_A)
     ref_P = ref_mp[:, 0]
@@ -492,8 +488,8 @@ def test_scrump_plus_plus_self_join(T_A, T_B, percentages):
                     ref_mp[i, 1] = ref_I[i]
             ref_P = ref_mp[:, 0]
             ref_I = ref_mp[:, 1]
-            ref_left_I = ref_mp[:, 2]
-            ref_right_I = ref_mp[:, 3]
+            # ref_left_I = ref_mp[:, 2]
+            # ref_right_I = ref_mp[:, 3]
 
             np.random.seed(seed)
             approx = scrump(
@@ -502,8 +498,8 @@ def test_scrump_plus_plus_self_join(T_A, T_B, percentages):
             approx.update()
             comp_P = approx.P_
             comp_I = approx.I_
-            comp_left_I = approx.left_I_
-            comp_right_I = approx.right_I_
+            # comp_left_I = approx.left_I_
+            # comp_right_I = approx.right_I_
 
             naive.replace_inf(ref_P)
             naive.replace_inf(comp_I)
@@ -698,9 +694,9 @@ def test_scrump_identical_subsequence_self_join(percentages):
         np.random.seed(seed)
         ref_mp = naive_scrump(T, m, T, percentage, zone, False, None)
         ref_P = ref_mp[:, 0]
-        ref_I = ref_mp[:, 1]
-        ref_left_I = ref_mp[:, 2]
-        ref_right_I = ref_mp[:, 3]
+        # ref_I = ref_mp[:, 1]
+        # ref_left_I = ref_mp[:, 2]
+        # ref_right_I = ref_mp[:, 3]
 
         np.random.seed(seed)
         approx = scrump(
@@ -708,9 +704,9 @@ def test_scrump_identical_subsequence_self_join(percentages):
         )
         approx.update()
         comp_P = approx.P_
-        comp_I = approx.I_
-        comp_left_I = approx.left_I_
-        comp_right_I = approx.right_I_
+        # comp_I = approx.I_
+        # comp_left_I = approx.left_I_
+        # comp_right_I = approx.right_I_
 
         naive.replace_inf(ref_P)
         naive.replace_inf(comp_P)
