+++ tests/test_stumpi.py
@@ -77,7 +77,6 @@ def test_stumpi_self_join():
 
 def test_stumpi_self_join_egress():
     m = 3
-    zone = int(np.ceil(m / 4))
 
     seed = np.random.randint(100000)
     np.random.seed(seed)
@@ -245,7 +244,6 @@ def test_stumpi_init_nan_inf_self_join(substitute, substitution_locations):
 @pytest.mark.parametrize("substitution_locations", substitution_locations)
 def test_stumpi_init_nan_inf_self_join_egress(substitute, substitution_locations):
     m = 3
-    zone = int(np.ceil(m / 4))
 
     seed = np.random.randint(100000)
     # seed = 58638
@@ -420,7 +418,6 @@ def test_stumpi_stream_nan_inf_self_join(substitute, substitution_locations):
 @pytest.mark.parametrize("substitution_locations", substitution_locations)
 def test_stumpi_stream_nan_inf_self_join_egress(substitute, substitution_locations):
     m = 3
-    zone = int(np.ceil(m / 4))
 
     seed = np.random.randint(100000)
 
@@ -547,11 +544,11 @@ def test_stumpi_constant_subsequence_self_join():
         stream.update(t)
 
     comp_P = stream.P_
-    comp_I = stream.I_
+    # comp_I = stream.I_
 
     ref_mp = naive.stamp(stream.T_, m, exclusion_zone=zone)
     ref_P = ref_mp[:, 0]
-    ref_I = ref_mp[:, 1]
+    # ref_I = ref_mp[:, 1]
 
     naive.replace_inf(ref_P)
     naive.replace_inf(comp_P)
@@ -568,7 +565,7 @@ def test_stumpi_constant_subsequence_self_join():
         stream.update(t)
 
     comp_P = stream.P_
-    comp_I = stream.I_
+    # comp_I = stream.I_
 
     naive.replace_inf(comp_P)
 
@@ -578,7 +575,6 @@ def test_stumpi_constant_subsequence_self_join():
 
 def test_stumpi_constant_subsequence_self_join_egress():
     m = 3
-    zone = int(np.ceil(m / 4))
 
     seed = np.random.randint(100000)
     np.random.seed(seed)
@@ -587,16 +583,16 @@ def test_stumpi_constant_subsequence_self_join_egress():
 
     ref_mp = naive.stumpi_egress(T, m)
     ref_P = ref_mp.P_.copy()
-    ref_I = ref_mp.I_
+    # ref_I = ref_mp.I_
     ref_left_P = ref_mp.left_P_.copy()
-    ref_left_I = ref_mp.left_I_
+    # ref_left_I = ref_mp.left_I_
 
     stream = stumpi(T, m, egress=True)
 
     comp_P = stream.P_.copy()
-    comp_I = stream.I_
+    # comp_I = stream.I_
     comp_left_P = stream.left_P_.copy()
-    comp_left_I = stream.left_I_
+    # comp_left_I = stream.left_I_
 
     naive.replace_inf(ref_P)
     naive.replace_inf(ref_left_P)
@@ -614,14 +610,14 @@ def test_stumpi_constant_subsequence_self_join_egress():
         stream.update(t)
 
         comp_P = stream.P_.copy()
-        comp_I = stream.I_
+        # comp_I = stream.I_
         comp_left_P = stream.left_P_.copy()
-        comp_left_I = stream.left_I_
+        # comp_left_I = stream.left_I_
 
         ref_P = ref_mp.P_.copy()
-        ref_I = ref_mp.I_
+        # ref_I = ref_mp.I_
         ref_left_P = ref_mp.left_P_.copy()
-        ref_left_I = ref_mp.left_I_
+        # ref_left_I = ref_mp.left_I_
 
         naive.replace_inf(ref_P)
         naive.replace_inf(ref_left_P)
@@ -639,16 +635,16 @@ def test_stumpi_constant_subsequence_self_join_egress():
 
     ref_mp = naive.stumpi_egress(T, m)
     ref_P = ref_mp.P_.copy()
-    ref_I = ref_mp.I_
+    # ref_I = ref_mp.I_
     ref_left_P = ref_mp.left_P_.copy()
-    ref_left_I = ref_mp.left_I_
+    # ref_left_I = ref_mp.left_I_
 
     stream = stumpi(T, m, egress=True)
 
     comp_P = stream.P_.copy()
-    comp_I = stream.I_
+    # comp_I = stream.I_
     comp_left_P = stream.left_P_.copy()
-    comp_left_I = stream.left_I_
+    # comp_left_I = stream.left_I_
 
     naive.replace_inf(ref_P)
     naive.replace_inf(ref_left_P)
@@ -666,14 +662,14 @@ def test_stumpi_constant_subsequence_self_join_egress():
         stream.update(t)
 
         comp_P = stream.P_.copy()
-        comp_I = stream.I_
+        # comp_I = stream.I_
         comp_left_P = stream.left_P_.copy()
-        comp_left_I = stream.left_I_
+        # comp_left_I = stream.left_I_
 
         ref_P = ref_mp.P_.copy()
-        ref_I = ref_mp.I_
+        # ref_I = ref_mp.I_
         ref_left_P = ref_mp.left_P_.copy()
-        ref_left_I = ref_mp.left_I_
+        # ref_left_I = ref_mp.left_I_
 
         naive.replace_inf(ref_P)
         naive.replace_inf(ref_left_P)
@@ -703,11 +699,11 @@ def test_stumpi_identical_subsequence_self_join():
         stream.update(t)
 
     comp_P = stream.P_
-    comp_I = stream.I_
+    # comp_I = stream.I_
 
     ref_mp = naive.stamp(stream.T_, m, exclusion_zone=zone)
     ref_P = ref_mp[:, 0]
-    ref_I = ref_mp[:, 1]
+    # ref_I = ref_mp[:, 1]
 
     naive.replace_inf(ref_P)
     naive.replace_inf(comp_P)
@@ -727,7 +723,7 @@ def test_stumpi_identical_subsequence_self_join():
         stream.update(t)
 
     comp_P = stream.P_
-    comp_I = stream.I_
+    # comp_I = stream.I_
 
     naive.replace_inf(comp_P)
 
@@ -737,7 +733,6 @@ def test_stumpi_identical_subsequence_self_join():
 
 def test_stumpi_identical_subsequence_self_join_egress():
     m = 3
-    zone = int(np.ceil(m / 4))
 
     seed = np.random.randint(100000)
     np.random.seed(seed)
@@ -749,16 +744,16 @@ def test_stumpi_identical_subsequence_self_join_egress():
 
     ref_mp = naive.stumpi_egress(T, m)
     ref_P = ref_mp.P_.copy()
-    ref_I = ref_mp.I_
+    # ref_I = ref_mp.I_
     ref_left_P = ref_mp.left_P_.copy()
-    ref_left_I = ref_mp.left_I_
+    # ref_left_I = ref_mp.left_I_
 
     stream = stumpi(T, m, egress=True)
 
     comp_P = stream.P_.copy()
-    comp_I = stream.I_
+    # comp_I = stream.I_
     comp_left_P = stream.left_P_.copy()
-    comp_left_I = stream.left_I_
+    # comp_left_I = stream.left_I_
 
     naive.replace_inf(ref_P)
     naive.replace_inf(ref_left_P)
@@ -778,14 +773,14 @@ def test_stumpi_identical_subsequence_self_join_egress():
         stream.update(t)
 
         comp_P = stream.P_.copy()
-        comp_I = stream.I_
+        # comp_I = stream.I_
         comp_left_P = stream.left_P_.copy()
-        comp_left_I = stream.left_I_
+        # comp_left_I = stream.left_I_
 
         ref_P = ref_mp.P_.copy()
-        ref_I = ref_mp.I_
+        # ref_I = ref_mp.I_
         ref_left_P = ref_mp.left_P_.copy()
-        ref_left_I = ref_mp.left_I_
+        # ref_left_I = ref_mp.left_I_
 
         naive.replace_inf(ref_P)
         naive.replace_inf(ref_left_P)
@@ -807,16 +802,16 @@ def test_stumpi_identical_subsequence_self_join_egress():
     T = pd.Series(T)
     ref_mp = naive.stumpi_egress(T, m)
     ref_P = ref_mp.P_.copy()
-    ref_I = ref_mp.I_
+    # ref_I = ref_mp.I_
     ref_left_P = ref_mp.left_P_.copy()
-    ref_left_I = ref_mp.left_I_
+    # ref_left_I = ref_mp.left_I_
 
     stream = stumpi(T, m, egress=True)
 
     comp_P = stream.P_.copy()
-    comp_I = stream.I_
+    # comp_I = stream.I_
     comp_left_P = stream.left_P_.copy()
-    comp_left_I = stream.left_I_
+    # comp_left_I = stream.left_I_
 
     naive.replace_inf(ref_P)
     naive.replace_inf(ref_left_P)
@@ -836,14 +831,14 @@ def test_stumpi_identical_subsequence_self_join_egress():
         stream.update(t)
 
         comp_P = stream.P_.copy()
-        comp_I = stream.I_
+        # comp_I = stream.I_
         comp_left_P = stream.left_P_.copy()
-        comp_left_I = stream.left_I_
+        # comp_left_I = stream.left_I_
 
         ref_P = ref_mp.P_.copy()
-        ref_I = ref_mp.I_
+        # ref_I = ref_mp.I_
         ref_left_P = ref_mp.left_P_.copy()
-        ref_left_I = ref_mp.left_I_
+        # ref_left_I = ref_mp.left_I_
 
         naive.replace_inf(ref_P)
         naive.replace_inf(ref_left_P)
